refactor(web): extract address inputs into utility
