+++ appeals/web/src/server/appeals/appeal-details/appellant-case/address/address.mapper.js
@@ -1,9 +1,5 @@
 import { appealShortReference } from '#lib/appeals-formatter.js';
-import {
-	errorAddressLine1,
-	errorPostcode,
-	errorTown
-} from '#lib/error-handlers/change-screen-error-handlers.js';
+import { addressInputs } from '#lib/page-components/address.js';
 
 /**
  * @typedef {import('../../appeal-details.types.js').WebAppeal} Appeal
@@ -18,89 +14,14 @@ import {
  */
 export function changeSiteAddressPage(appealData, backLinkUrl, sessionData, errors) {
 	const shortAppealReference = appealShortReference(appealData.appealReference);
-	const siteAddress = sessionData ?? appealData.appealSite;
+	const address = sessionData ?? appealData.appealSite;
 
-	/** @type {PageContent} */
-	const pageContent = {
+	return {
 		title: `Change site address`,
 		backLinkUrl: backLinkUrl,
 		preHeading: `Appeal ${shortAppealReference}`,
 		heading: 'Change site address',
 		headingClasses: 'govuk-heading-l',
-		pageComponents: [
-			{
-				type: 'input',
-				parameters: {
-					id: 'address-line-1',
-					name: 'addressLine1',
-					type: 'text',
-					label: {
-						isPageHeading: false,
-						text: 'Address line 1'
-					},
-					value: siteAddress?.addressLine1 ?? '',
-					errorMessage: errorAddressLine1(errors)
-				}
-			},
-			{
-				type: 'input',
-				parameters: {
-					id: 'address-line-2',
-					name: 'addressLine2',
-					type: 'text',
-					label: {
-						isPageHeading: false,
-						text: 'Address line 2 (optional)'
-					},
-					value: siteAddress?.addressLine2 ?? ''
-				}
-			},
-			{
-				type: 'input',
-				parameters: {
-					id: 'address-town',
-					name: 'town',
-					type: 'text',
-					classes: 'govuk-input govuk-input--width-20',
-					label: {
-						isPageHeading: false,
-						text: 'Town or city'
-					},
-					value: siteAddress?.town ?? '',
-					errorMessage: errorTown(errors)
-				}
-			},
-			{
-				type: 'input',
-				parameters: {
-					id: 'address-county',
-					name: 'county',
-					type: 'text',
-					classes: 'govuk-input govuk-input--width-20',
-					label: {
-						isPageHeading: false,
-						text: 'County (optional)'
-					},
-					value: siteAddress?.county ?? ''
-				}
-			},
-			{
-				type: 'input',
-				parameters: {
-					id: 'address-postcode',
-					name: 'postCode',
-					type: 'text',
-					classes: 'govuk-input govuk-input--width-10',
-					label: {
-						isPageHeading: false,
-						text: 'Postcode'
-					},
-					value: siteAddress?.postCode ?? '',
-					errorMessage: errorPostcode(errors)
-				}
-			}
-		]
+		pageComponents: addressInputs({ address, errors })
 	};
-
-	return pageContent;
 }
